+++ src/utils/eventBasedTracking/viewSalaryWorkTime.js
@@ -1,4 +1,4 @@
-import { getEventBasedTrackingClient } from './client';
+import { getEventBasedTrackingClient } from './__client';
 
 export const sendEvent = ({
   company,
